% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wap.R
\name{projM}
\alias{projM}
\title{Projection Matrix}
\usage{
projM(X, ...)
}
\arguments{
\item{X}{matrix with dimensions \eqn{n \times p}.}

\item{...}{additional matrix or vectors that have same column 
dimension with X.}
}
\value{
A diagonal matrix.
}
\description{
Calculate the projection of matrix
}
\examples{

\dontrun{
  X=matrix(10,rnorm(100))
  projM(X)
  
  # add additional vector Y
  Y=rep(1,10)
  projM(X,Y) 
}

}
