% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wap.R
\name{wpvalue}
\alias{wpvalue}
\title{Weighted p-value}
\usage{
wpvalue(Y, X, Z, V = NULL, subset = NULL, Raw.Weight = FALSE,
  method = c("max", "mean", "Lw"), ncores = 1, center = TRUE,
  scale = FALSE)
}
\arguments{
\item{Y}{A vector of numerical values. response variable.}

\item{X}{A design matrix.}

\item{Z}{A matrix for intermediate variables.}

\item{V}{A matrix for external variables. Default is NULL.}

\item{subset}{A vector of indices for a sub-sample of the data or a subsample size. 
subset is TRUE for method ``Lw". Default is NULL.}

\item{Raw.Weight}{Whether or not return the raw weight. Default is FALSE.}

\item{method}{weight methods. Must choose one from ``mean",``max", and ``Lw".}

\item{ncores}{number of cpus. See \code{\link[parallel]{makeCluster}}.}

\item{center}{Whether or not centerize the data. Default is FALSE.}

\item{scale}{Whether or not scale the data. Default is FALSE.}
}
\value{
A dataframe consists of raw p-values, weight, and adjusted p-values. When the Raw.Weight
        is TRUE, a list will be returned with the result dataframe and the raw weight.
}
\description{
Calcuate adjusted p-value by using weighted p-value approach
}
\examples{

\dontrun{
X=matrix(rnorm(500),50)
beta1=0.5;beta2=0.1
Z1=X[,1]*beta1+rnorm(50)
Y=Z1*beta2+rnorm(50)
Z=cbind(Z1,matrix(rnorm(50000),50))
test=wpvalue(Y,X,Z,method="max",ncores=4)
} 

}
